Crate tidb_query_aggr

This crate provides aggregate functions for batch executors.

Modules

impl_avg

impl_bit_op

impl_count

impl_first

impl_max_min

impl_sum

impl_variance

parser

summable

util

Macros

impl_concrete_state

impl_state_update_partial

impl_unmatched_function_state

update

update_concrete

update_repeat

update_vector

Structs

AllAggrDefinitionParser
Parse all aggregate function definition from protobuf.

Traits

AggrDefinitionParser
Parse a specific aggregate function definition from protobuf.

AggrFunction
A trait for all single parameter aggregate functions.

AggrFunctionState
A trait for all single parameter aggregate function states.

AggrFunctionStateUpdatePartial
A helper trait that provides update() and update_vector() over a concrete type, which will be relied in AggrFunctionState.

ConcreteAggrFunctionState
A helper trait for single parameter aggregate function states that only work over concrete eval types. This is the actual and only trait that normal aggregate function states will implement.