Trait tidb_query_aggr::AggrFunction

pub trait AggrFunction: Debug + Send + 'static {
    fn name(&self) -> &'static str;

    fn create_state(&self) -> Box<dyn AggrFunctionState>;
}

A trait for all single parameter aggregate functions.

Unlike ordinary function, aggregate function calculates a summary value over multiple rows. To save memory, this functionality is provided via an incremental update model:

Each aggregate function associates a state structure, storing partially computed aggregate results.

The caller calls update() or update_vector() for each row to update the state.

The caller finally calls push_result() to aggregate a summary value and push it into the given data container.

This trait can be auto derived by using tidb_query_codegen::AggrFunction.

Required methods

fn name(&self) -> &'static str

The display name of the function.

fn create_state(&self) -> Box<dyn AggrFunctionState>

Creates a new state instance. Different states aggregate independently.

Loading content...