Trait tidb_query_aggr::ConcreteAggrFunctionState

pub trait ConcreteAggrFunctionState: Debug + Send + 'static {
    type ParameterType: EvaluableRef<'static>;
    unsafe fn update_concrete_unsafe(
        &mut self, 
        ctx: &mut EvalContext, 
        value: Option<Self::ParameterType>
    ) -> Result<()>;

    fn push_result(
        &self, 
        ctx: &mut EvalContext, 
        target: &mut [VectorValue]
    ) -> Result<()>;
}

A helper trait for single parameter aggregate function states that only work over concrete eval types. This is the actual and only trait that normal aggregate function states will implement.

Unlike AggrFunctionState, this trait only provides specialized update() and push_result() functions according to the associated type. update() and push_result() functions that accept any eval types (but will panic when eval type does not match expectation) will be generated via implementations over this trait.

Associated Types

type ParameterType: EvaluableRef<'static>

Loading content...
Required methods

unsafe fn update_concrete_unsafe(
&mut self,
ctx: &mut EvalContext,
value: Option<Self::ParameterType>
) -> Result<()>

Safety

This function should be called with update_concrete macro.

fn push_result(
&self,
ctx: &mut EvalContext,
target: &mut [VectorValue]
) -> Result<()>

Loading content...